#518 - remaining multi tenant ooperations

+++ src/database/defaultSettings.ts
@@ -64,7 +64,7 @@ export async function deleteDefaultSetting(id: number, tenantId: string): Promis
 
     let result;
     try {
-        [result] = await getPool().query<OkPacket>('delete from default_settings where default_setting_id = ?', values);
+        [result] = await getPool().query<OkPacket>('delete from default_settings where default_setting_id = ? and tenant_id = ?', values);
     } catch (e) {
         logger.error(`DB error deleting default setting: ${e}`);
         throw new Error('internal server error');
